Use non-ambiguous variable names in PL/pgSQL code
In particular, the 'directory' variable used in filepath(),
directorypath(), subdirectories() and containedfiles() conflicted
with the 'directory' column in the 'directories' table, used in
queries in the functions.  Treating this as an error is apparently
new in PostgreSQL 9.0.

Rename the variable 'directory_out' or 'directory_id' instead.
Also rename the variable 'file' in containedfiles() 'file_out',
for consistency.