Avoid adding duplicated review filters on /createreview
If the user used the "Add Reviewer" or "Add Watcher" buttons to make
the same user a reviewer or watcher of the same path multiple times,
duplicate entries would be submitted to the server when submitting the
review, which caused the submit to fail.

Instead, just silently replace conflicting entries when new are added.

(The review filters UI on /createreview is rather bad and could use a
thorough redesign.)