Make /createreview a bit more flexible
Support specifying an arbitrary commit as the upstream to use when
selecting commits to include in the review instead of requiring it
to be a branch in the remote repository.  There might not be a
suitable branch in the remote repository, and it's silly to require
that one is created there just to create a review.

If the upstream is on the form "refs/*" it is fetched from the
remote repository, otherwise it is simply rev-parse:d in Critic's
repository after the work branch has been fetched.  This means an
abbreviated SHA-1 will work as long as the specified commit is
reachable from the head of the work branch and thus was fetched
when along with the work branch.