Give useful error message when creating comment via /showfile
If the commented file has been changed in the review (in any way
other than just being added or removed) the code for creating
comments can't deal with creating a comment in a single version
of the file rather than in the context of a diff.  Give a useful
error message when this limitation is hit, including a link to
a diff via which the comment could be created instead.

This is of course really just a stupid limitation that should be
fixed, but until then...