Consider both .cc and .cpp as source files when sorting paths
When displaying diffs, paths are sorted with header files preceding
source files (when the paths minus the extension are identical) since
it's typically more convenient to see the changes in a header before
the changes in the corresponding source file.  This comparison used
to consider only *.cpp as source files and only *.h as header files.

This should probably be made customizable or otherwise more flexible.