Ask for SMTP configuration during installation
Used to be hardcoded to use localhost:25, no authentication and no TLS.
Also offer to send a test email to verify the configuration.

Also add support using an SSL/TLS connection to the SMTP server (using
smtplib.SMTP_SSL() instead of smtplib.SMTP()) as an alternative to
using STARTTLS to upgrade a plain-text connection.