Support (restricted) anonymous access
If Critic authenticates users and keeps track of session using cookies,
it's obviously possible to defer sign in until the user explicitly
decides to, or has to, and provide restricted anonymous access until
then.

Pages by default allow anonymous access, with possible opt-out by
redirecting to the /login page.  Operations, on the other hand, are by
default limited to signed in users, with possible opt-in to allow
anonymous access.