Fix compacting of syntax highlighting cache
Crashed if the cache directory was missing, which it is in a newly
installed system.

Make time of day to run compacting configurable.  In particular useful
in settings where multiple instances share the same cache, in which case
it's now possible to disable the compacting in all but one instance.

Add support for manually triggering compacting by sending a command to
the background process over its command socket.  (No UI yet.)

Also fix some minor bugs that caused the compacting to not compact.

Configurable time of day and manual trigger also applies to the purging
of older custom changesets from the database.