Fix display of "watched" category without "active" category
On /dashboard, if the "watched" category was enabled but not the "active"
category, fetchActive() was called between the query for watched reviews
and the processing of the resulting rows.  But fetchActive() used the same
cursor object to make its own query and thus trampled the result of the
query for watched reviews.

(The list of active reviews needs to be fetched so that the active reviews
can be excluded from the list of watched (inactive) reviews where they
don't belong.)