Sanitize operation arguments value in error messages
If /validatelogin or /changepassword were to fail unexpectedly, the
automatic error message sent to the system administrator would contain
the user's password in plain text.  That's not a great idea, so allow
an operation to sanitize its arguments value before it's used in error
messages (or, in a hypothetical future, in logs.)

Also fix handling of errors (other than the basic "wrong username or
password" failure) on /login; the form submit wasn't cancelled, so the
user would be redirected to the anonymous user's default page instead
of staying at the /login page.