Fix commenting via "detect moved code" diff
The diff for detected moves consists of dummy diff.File objects with zero
file id, and references to the real source and target diff.File objects
for the move instead.  When generating the HTML code, however, no special
handling of this existed, so the id attributes on the table cells, which
are used by the client-side scripts when creating comments to figure out
what lines in what file was selected, used the incorrect zero file id.