Fail less hard if an installed extension is missing
For inject roles, extension errors should be handled by inserting an
HTML comment with an error message into the target page, rather than
by failing completely to load the target page.  This was done for most
types of extension errors, but not if the extension's manifest failed
to load, for instance because the extension had been removed.