Put single class on the row in addition to column.
Makes it possible to target and apply specific styling in case of
one diff pane view (newly added file for example). It was impossible
to apply some specific styling in this case. For example making
gutter on both sides have green background but have it only on the
right in 2-pane view was more or less impossible.