Fix upgrading that removes source files
The SHA-1 of the existing file was calculated incorrectly (Git
doesn't just hash the contents of the file) so when comparing to
the file's SHA-1 according to Git, all files appeared to have been
modified.

Also, if a file to remove wasn't detected as having been changed
(which before this fix never happened) then it wouldn't be removed
at all, since removing it was strictly conditioned on the user
answering "remove it" when asked what to do.

Also, if a Python source file A.py is removed, remove A.pyc and
A.pyo as well.  If they are left, Python is likely to pick them up
and use them even if the original A.py is gone.