Add debugging code for incorrectly re-used cursor objects
If a cursor object is currently being iterated over, it's a bad idea
to execute a new statement using it.  What's worse, the database query
profiling preference hid the problems caused by doing so.

With this code (which is only used when query profiling is enabled)
incorrectly re-using a cursor object while iterating over it triggers
an exception the next time next() is called on the iterator.