Remove dbutils.Review.getCommentChains()
The function was called from a single location, and had two modes, one
in which it simply called CommentChain.fromReview() and one in which it
ran a database query and created CommentChain objects itself.

Only the former mode was ever used so the single location in which this
function was used now calls CommentChain.fromReview() directly instead.