Preserve file mode information in reduced merge changeset
Reduced/filtered merge changesets copy files from the full diff between
parent and merge commit, with a filtered list of chunks.  The file mode
information was lost in this copying, which is not an advantage to begin
with, and is really bad for sub-module changes, since the fact that it is
a sub-module change is indicated only by the file mode.  Rendering a
change in a sub-module (git link) as a change in a regular file fails,
since the SHA-1 of the tree entry typically isn't present in the
repository (since it's the SHA-1 of a commit in another repository.)