Remove broken /showfile link for deleted files
The /showfile link from diffs always link to the new version of the file.
If the file was deleted, there is no new version, and the link is broken.
It would be meaningful to link to the old version of the file instead in
this case, but as a workaround until that's fixed, don't link at all.