Improve parsing of irregular diffs
If a file was changed in a complex way, for instance from a symbolic link
to a binary file, the diff parser was confused.  In such cases, the diff
will be that one file was removed and another, by the same name, was
independently added.  This was handled if it was added as a regular file,
but not if it was added as a binary file.