Process commits when tags are created or updated
Processing commits means recording the commits themselves in the database
in the 'commits' table.  This wasn't done when tags were pushed, and
wasn't immediately required, since the 'tags' table doesn't actually have
a foreign key to the 'commits' table (it just stores the SHA-1.)  But
having tags in the 'tags' table referencing the SHA-1 of commits that are
in the repository, but not in the 'commits' table, is just wrong.