Fix raising issues in separately fetched removed or added file
A removed file, or a very large added file, is not included when diffs
are rendered.  Instead they are displayed with a button allowing the
user to fetch and display the contents of the file.  The HTML generated
when inserting the contents into the page was not entirely correct, so
commenting the fetch lines did not work.