Improve handling of diffs between non-linear commits
When diffing between two commits and filtering to only include changes in
certain files, we can either do a full diff, which gets cached in the DB,
or do a "thin" diff where we run "git diff" directly for only the filtered
files, and don't store the result in the DB.

The heuristics for when to do the latter was if there was a merge commit
"between" the two commits.  However, the loop used to detect that assumed
that one commit was an ancestor of the other, which might not be the case.
Add check for that, and if it is not, do the latter.