Make "conflicts" diff viewing a bit more flexible
In addition to the old mode where a single commit (always a merge) is
displayed in "conflicts" mode, a new mode where a diff between two commits
is displayed in "conflicts" mode is added.  In this mode, the first commit
is assumed to have been produced by replaying a rebase or a merge, and the
second commit is assumed to be the actual result of the rebase or merge.