Fix installation failure if PostgreSQL doesn't speak English
As part of the database setup, 'createlang' is run to enable/install
the PL/pgSQL language.  This command fails if the language is already
enabled (enabling it in one database causes it to be enabled in all
subsequently created databases automatically,) a failure we wish to
ignore since it doesn't affect the rest of the installation.

The specific failure "language already enabled" was ignored by
looking at the error message output by 'createlang', which in turn
fails if the error message wasn't in English.  Even if we ran
'createlang' with LANG=C or something similar that causes it to give
an error message in English, it is still unfortunate to rely on the
exact wording of an error message intended for human consumption.

Instead, ignore all failures from 'createlang'.  If the operation
actually failed, and the language isn't enabled, later steps in the
database setup will fail anyway, so there's no risk of producing an
incomplete or broken install.