README: Expand on steps needed after the initial installation
Explain how to set up SSH for the "critic" user so that it can access
upstream repositories properly.

Add a mention of the "criticctl" tool, which is used to add users to the
Critic user database.

Change "account" to "system account" under "Adding push rights", to
differentiate with the Critic user account mentioned in the paragraph
just above.