Add table for keeping track of remotes that "pushes" updates
If a remote has a post-update hook that talks to the branchtrackerhook
service, it's rather unnecessary to poll for updates to tracked branches
in that remote every hour.  (Or at all, really.)

With this fix, tracked branches in any remote listed in the 'knownremotes'
table (with the 'pushing' flag/column set to TRUE) are polled for updates
once a week instead of once an hour.  Remotes are added to the table
automatically whenever they actually push updates.