Don't run migration scripts that existed when Critic was installed
For most migration scripts this doesn't matter; they are (and will
continue to be) written to be harmless or no-ops if run when they don't
need to be run.

But for migration scripts that add news items it does matter.  Without
this change the first time Critic is upgraded after installation, all
existing migration scripts would run, and all migration scripts that
add news items would add their news items, despite the news being old.

This is not the end of the world, but is a bit unexpected.

Running DB schema migration scripts also triggers the warning about
the DB schema being modified before the prompt asking whether to make
a dump of the database, and changes the default answer in that prompt
to true.  This is of course just unnecessary if the DB schema will in
fact not be modified.