Tune filter ranking slightly
When counting the number of forward slashes in a filter path, do not
include slashes that follow "**", since "**/" matches zero or more path
components, and thus might not actually represent a slash in the final
path.

This means "**/*.txt" is ranked below (less specific) than "foo.txt"
which intuitively makes sense.  Before this adjustment, it would have
been ranked above it on account of it containing more forward slashes.