Fix crash when rendering certain addressed issues
If an issue is marked as addressed by a non-fast-forward rebase, show
the original diff (with the comment on the right-hand side) instead of
the diff of the commit that addressed it.  The normal way of displaying
an addressed issue assumes that the issue is present in the left-hand
side of the (direct) diff of the commit that addressed the issue, but
this is usually not true in this particular case.