Fix OperationFailureMustLogin exception
Both its __init__() and __str__() incorrectly used

  super(OperationFailure, self)

to get the "base class instance" of 'self', but this returned instance
belonging to the base class of OperationFailure, which is Exception,
meaning it accidentally ended up bypassing its immediate base class.

Also, OperationFailureMustLogin.__str__() was completely redundant; it
meant to redirect to its immediate base class's __str__(), but the
easiest way to do that is of course to not override __str__() in the
first place.