Add --skip-testmail-check argument to install.py
This argument disables the "Did the test email arrive correctly?" question
after the test email has been (seemingly) successfully sent.  Useful when
installing as part of testing; the testing framework will catch the email
and check that it was correct, but doesn't want to have to tell the
installation script whether it arrived correctly.