Use temporary clones for relaying instead of temporary remotes
The branch tracker adds a temporary remote to the persistent relay
repository each time it wants to fetch a branch from a remote repository.
This works, and saves us the trouble of keeping track of whether there
is a suitable remote or not and what its name is (and some other special
cases.)  However, it turns out that not keeping the remote (and thus a
record of what refs have been fetched from it already) causes "git fetch"
to fetch a lot of objects over and over from the remote, despite them
already being in the local repository, which causes the local repository
to grow indefintely (until GC:ed or recreated.)

To avoid this, and also avoid potential race conditions from concurrent
use of the persistent relay repository, do the relaying in a temporary
clone instead.  Creating a bare clone on the same file system is almost
instantaneous and cost-free.

TODO: Should document the fact that because of this, in particular the
assumption that creating a clone is very cheap, DATA_DIR and GIT_DIR
should be located on the same file-system.  The install script should
also warn if the configured paths are not on the same file-system.