Improve support for deleting reviews
There is no UI for deleting reviews, but it can be done using a direct
connection to the database, and, in rare cases, makes sense to do.

To better support this, add "ON DELETE SET NULL" to some foreign key
constraints referencing the 'reviews' table, and add an index to the
'reviewmessageids' table (which can be rather huge) so that its "ON
DELETE CASCADE" foreign key constraint is less expensive to apply.