Support "automatic" rebase of reviews tracking remote branches
The user inputs the name of a new, rebased branch in the same remote as
the currently tracked branch.  Alternatively, if the customization
module customization.branches has a functions getRebasedBranchPattern()
and isRebasedBranchCandidate(), the remote can be scanned for likely
rebased branches based on some branch naming scheme, if there is one.

The rebased remote branch is fetched, its status as a rebased version of
the branch is checked, indicating whether there seem to be conflicts or
unexpected changes on it compared to the current review branch.  If all
looks well, the rebase can be performed (without the need to run any git
commands) and the review is set to track the new branch in the remote
instead.