Upgrade to jQuery 2.0.0 and jQuery UI 1.10.2
Also remove jquery-tooltip plugin and use the tooltip functionality that
is included in jQuery UI.  The API was slightly different so some code in
client-side scripts needed to be updated.  (The tooltips on comment
markers got a bit of a face-lift in the process.)

Also change code that manipulates the disabled state of form controls to
use

  obj.prop("disabled", true|false)

instead of

  obj.attr("disabled", "disabled")
  obj.removeAttr("disabled")

as suggested in the jQuery 1.9 upgrade guide.

Also add the style-sheet jquery-ui-overrides.css to override specific
styles in jquery-ui.css.  For now, it effectively reverts a change in
jQuery UI relative the previously used version that changes the font
used by widgets ("Verdana" had been removed from the list of fonts.)