Fix issue where direct pushes to tracking reviews were allowed
Direct pushes to review branches that track branches in other repositories
are supposed to be rejected.  But since the branch tracker service uses
"git push" to update branches too, an exception needs to be made.  The
exception was implemented by checking if the user was the Critic system
user.

This however is not conclusive.  Extensions, for instance, run as the
Critic system user, so "git push" executed by an extension would still be
allowed, incorrectly.

This commit improves the exception by adding support for passing "flags"
from the pushing Git process to the githook service via the $CRITIC_FLAGS
environment variable (valid only if the Critic system user is pushing.)
This mechanism is used to pass a flag saying that the push is used to
update a given tracked branch.