Improve detection of whether comments are active
The logic for determining whether a comment is active or not ("present in
the commit at the tip of the branch") used "commit has no children in the
set of relevant commits" as the criterion for determining this.  It is
however possible to achieve a sort of cycle in the commit-set, in which
case this is not true for any commit, and no comments can then be active.

One way to achieve such a cycle is to have a review containing the commits
A, B and C, where C reverts B, and then do a history rewrite rebase back
to A.  A is then the tip of the branch, but still has children in the
commit-set.  In this situation, a comment present in A should be active,
even if changes in B/C touch the lines.

Fix by changing the logic to actually explicitly detect that the comment
is present in the commit at the tip of the branch.  Sometimes this commit
can be found as "self.review.branch.head", but notable can't when commits
are added to a review, since "self.review.branch.head" won't have been
updated yet.  In the latter case, use the head of the set of added commits
instead.