Don't ignore empty extension pages
An extension page role can skip a request by not writing anything, in
which case extensions.role.page.execute() returns False, signalling that
the extension didn't handle the request.  But the extension can return
a response containing an empty response body, and that should not be
treated the same way; such an empty response should just be returned to
the client.