Improve handling of invalid upstream commit reference
If the upstream commit is not given as "refs/<something>", it's resolved
locally, and was assumed to be a valid commit.  It could either fail to
resolve as a reference, or it could be a full SHA-1 that just doesn't
name a commit in the local repository.  Handle these cases better.

In the latter case, since all possible valid upstream commits would have
been fetched along with the remote branch in the first place, the error
message points out that this must mean that the specified upstream commit
is in fact invalid.