Add possibility to opt out of review emails by default
If a user enables the preference 'review.defaultOptOut' then the user
automatically opts out of review emails (after the "New review" email)
in any review that isn't opt-in.

This is considered a short-term quick-fix solution; long term this
should be an option per filter rather than per user, but that requires
a larger set of changes.