Perform database updates later to avoid premature commit
The calls to reviewing.utils.addCommitsToReview() can (unfortunately)
cause the database transaction to be committed.  This means that any
updates done before those calls might be prematurely committed in cases
where a crash causes the push to be rejected, leading to the repository
and the database to become out-of-sync.

Avoid this by moving all database updates to after those calls.

Also drive-by-reformat code to restrict line-lengths.