Call added/deleted zero-length file "empty" instead of "binary"
Adding or deleting an empty file is recorded in the database in a form
that is indistinguishable from adding or deleting a binary file.  Whether
it makes sense to call a empty file "binary" or not is sort of undefined
(it is clearly unreasonable to call an empty "__init__.py" file binary,
whereas calling an empty "data.bin" or "invalid.jpg" file binary would
probably seem reasonable to most people.)  It always makes sense to call
an empty file "empty" though, regardless of whether it's an empty binary
file or an empty text file.