Clean diff.File objects a little less
With "Support non-fastforward move rebases" (12d0126) the same changeset
objects are used to render the diffs in each mail sent when commits are
added to a review.  This conflicts with changeset.text.unified()'s use of
diff.File.clean() to reduce peak memory usage, since that cleans too
aggressively.

Add a diff.File.cleanLines() that cleans only the lines, which is not
loaded by default anyway, and which changeset.text.unified() expects to
have to load itself already.