Move bulk of /usr/bin/criticctl to maintenance/criticctl.py
The part left is essentially that which depends on installation data,
and enough to make it possible to pull in maintenance/criticctl.py and
call the code in it.

Since templates are essentially large Python format strings, they are
cumbersome and error-prone to work with in the long run, and it makes
sense to keep them as small and simple as possible.