Call setgid('critic') in criticctl (already called setuid('critic'))
If /etc/critic is owned by root.critic, and not accessible to others, and
criticctl is called by root.root, which then setuid's to critic, and is
effectively critic.root, then /etc/critic is inaccessible to criticctl.

And running as critic.critic is the ideal anyway.

(Here, "critic" is the user/group that Critic was installed to run as.)