Add support for testing extensions
If the extensions/v8-jsshell sub-module is checked out, it is also
exported to the guest system (as git://host/v8-jsshell.git) thus
enabling it to be checked out there using "git submodule update".
(The extensions/v8-jsshell/v8 submodule is fetched from github.com.)

Extensions is enabled by a test (similar to how Critic is installed
and upgraded) that calls virtualbox.Instance.extend().  To improve
performance, the actual jsshell binary is cached (per instance and
extensions/v8-jsshell sub-module SHA-1) so that it doesn't need to
be built during every test run.  When a cached binary is found, the
"--fetch" and "--build" extend.py actions are skipped.

Extension testing is currently opt-in; the testing framework needs
to be started with --test-extensions for it to be enabled.  If not,
the virtualbox.Instance.extend() function raises a NotSupported
exception, which is like a TestFailed extension in that it causes
dependent tests to be skipped, but is not considered an error.