Redirect from HTTP to HTTPS when it makes sense
The /login page redirects to HTTPS if accessed over HTTP to avoid sending
credentials in plain text over an unencrypted connection.

The "sid" cookie set by the /validatelogin operation is set with the
"Secure" modifier, which (should) cause the user agent to only send the
cookie over secure connections.  To avoid getting an anonymous page when
loading an HTTP URL while logged in, an additional cookie, "has_sid", is
set (with a dummy value) which, when set, will cause all pages to
redirect from HTTP to HTTPS, so that the "sid" cookie gets sent.