Fix minor criticctl upgrading glitch
When criticctl is not modified by an upgrade (and hasn't been modified
since it was installed,) both of the "previously installed version matches
current version" and "current version matches to-be-installed version"
checks are true.

One leads to the file being updated (without actually modifying it) and
a message saying it was updated, the other leads to nothing at all.  It's
better to prefer to pick the one that leads to nothing at all, so shift
the two checks around.