Unbreak layout on /checkbranch (the input form version)
At some point the structure of the branch status version of the page and
stylesheet (checkbranch.css) was updated, but the structure of the input
form version of the page was not given matching love.  Do that now, and
also "modernize" it somewhat.

Also, some minor tweaks to basic page generation utilities:

- Always set class ("repository") on the <select> element generated by
  generateRepositorySelect(), and add basic styling of it to basic.css.

- Let PaleYellowTable set document title to same as table title if the
  document doesn't already have a title set.