Read output from guest commands more carefully
It seems readline() doesn't work reliably when the underlying file is
non-blocking.

Note: This code does a lot of string concatenations and is thus far from
optimized.  This is not considered a problem; this is just the testing
framework, after all, and performance is unlikely to suffer that much
anyway.  (Besides, the old code did something quite similar.)