Make repository.workcopy() somewhat more flexible
Support creating an empty repository (no content, no remotes.)  Mostly a
useless feature, but convenient if the purpose is to create an "initial"
commit.

Also support running git with custom environment variables; any upper-case
keyword argument to workcopy.run() is set as an environment variable for
the git process instead of being passed to subprocess.Popen().