Prepend install_dir to sys.path while installing
This way, when an installation script imports Critic modules it
imports them from install_dir instead of from $PWD.  This makes a
difference if the installation script has switched uid/gid to the
Critic system user/group, in which case it might not have access
to files in $PWD anymore.

Specifically, this makes it possible for installation/prefs.py to
import dbutils.py after setting uid/gid.