Start Critic background services before Apache
The only real upstart dependency is that wsgistartup.py tries to
write its pid file to a directory created by the init.d script and
fails if the init.d script hasn't created the directory yet.

That problem could obviously be fixed in wsgistartup.py as well.

But technically the web front-end does depend on the background
services, so ought to start after.