Support link types with leading non-word characters better
Requiring a word-boundary before a link "word" means that link types with
leading non-word characters must be immediately preceded by a word
character, which is unlikely to be the case.

Relax the requirement by requiring either a word boundary before the link
"word" or that the link "word" begins with a non-word character.  (Such
a relaxation was already in effect, but only for a select few non-word
characters.)