Improve handling of rebases in branch log renderer
Make use of the fact that we can sort rebases chronologically by their id
number, instead of trying to order them implicitly by linking their old
and new head commits together with other commits.  The latter fails in
various cases where a commit is the new head of multiple rebases, or when
a rebase simply resets the branch back to a previous point (such as when
the last commit on the branch is reverted, and it and the revert commit is
later removed by a history rewrite rebase.)

Also represent rebases as objects (of a local class) with named properties
instead of as tuples, for increased readability, and possible future world
domination.