Use uid=NULL instead of uid=zero to represent recipient list default
The uid column has a foreign key constraint to the users table, and thus
using zero requires the existance of a (dummy) user with id zero.  Using
NULL is compatible with the foreign key constraint (i.e. the constraint
is not enforced if the uid column is NULL.)

However, the table's primary key needs to be changed into a unique
constraint, since a primary key doesn't allow NULL values.