Add support for "variant" operation arguments
Declared as a set() of accepted types.  This overloads the existing use
of set() (of strings) for "enumerated values", but compatibly so.  Before
this change, any set() not containing only strings would have triggered
an error, and after this change a set() containing only strings work the
same way as before.