Support per-repository and per-filter preferences
Support different default settings per repository, and different user
overrides per repository and per filter.  Not all preferences can be set
per repository or filter; this is indicated by flags in the 'preferences'
table.  (Essentially this depends on whether the code that reads the
preference passes in the relevant repository or filter parameters, and
will be determined and implemented per setting.)

Also "modernize" the UI on /config.  Scrap the "Save Settings" button in
favor of automatic saving triggered by changes.

Also add support for editing system defaults and repository defaults,
available to users with the administrator role.

Also change how system default values are stored.  Instead of storing
those values in columns in the 'preferences' table, store them as rows in
the 'userpreferences' table with uid=NULL.  (The 'userpreferences' table's
name is by now quite inaccurate.)