Make review.defaultOptOut configurable per repository and filter
Policy (also documented in the code):

If all of the user's filters that matched files in the review have
review.defaultOptOut enabled, then opt out.  When determining this, any
review filters of the user's that match files in the review count as
filters that don't have the review.defaultOptOut enabled.

If any of the user's filters that matched files in the review have
review.defaultOptOut disabled, then don't opt out.  When determining
this, review filters are ignored.

Otherwise, ignore the filter settings, and go with either the user's
per-repository or global setting (as set above.)