Improve handling of non-ascii path names
This is not really a good solution.  Long-term we should reduce our
dependency on the "human-readable" aspects of the output from "git diff".
The particular "path munging" that we undo here could be disabled using
the -z argument to "git diff", but that only works together with --raw
(and some other flags) so requires some other changes as well.