Fix crash when performing non-ff rebase
The crash happened when propagating comments when adding the "replayed
rebase" commit to the review, whenever any commit was not addressed by
the changes.

The fix consists of three parts:

 1) Store the (just now) performed rebase explicitly in the review object
    so that it is taken into account properly when propagating comments.

 2) Use the "replayed rebases" mapping to set the correct head commit when
    propagating comments.  (The mapping maps the SHA-1 of the replayed
    rebase commit to the actual new head after the rebase.)

The first change became necessary after the changes in e73afd450c which
delayed the update of rebase's row in the 'reviewrebases' table.